+++ src/app/static/src/tests/components/errorReport.test.ts
@@ -14,6 +14,7 @@ import ErrorAlert from "../../app/components/ErrorAlert.vue";
 import {RootState} from "../../app/root";
 import { ErrorsState } from "../../app/store/errors/errors";
 import LoadingSpinner from "../../app/components/LoadingSpinner.vue";
+import {emailRegex} from "../../app/utils";
 
 describe("Error report component", () => {
 
@@ -485,6 +486,114 @@ describe("Error report component", () => {
         expect(wrapper.vm.$data.stepsToReproduce).toBe("");
         expect(wrapper.vm.$data.section).toBe("");
         expect(wrapper.vm.$data.email).toBe("");
+    });
+
+    it("can validate email with whitespace", async () => {
+        const store = createStore({}, {}, {}, true)
+        const wrapper = mount(ErrorReport, {
+            propsData: {
+                open: true
+            },
+            store
+        });
+
+        wrapper.find("#email").setValue("tes t@email.com ");
+        expect(wrapper.vm.$data.email).toBe("test@email.com");
+        expect(wrapper.find(".invalid-feedback").exists()).toBe(false)
+    });
+
+    it("does not validate email with one letter postfix", async () => {
+        const store = createStore({}, {}, {}, true)
+        const wrapper = mount(ErrorReport, {
+            propsData: {
+                open: true
+            },
+            store
+        });
+
+        wrapper.find("#description").setValue("something");
+        wrapper.find("#reproduce").setValue("reproduce steps");
+        wrapper.find("#section").setValue("downloadResults");
+        wrapper.find("#email").setValue("test@email.c");
+
+        expect(wrapper.find(".btn-red").text()).toBe("Send");
+        expect(wrapper.find(".btn-red").attributes("disabled")).toBe("disabled")
+
+        const invalidFeedback = wrapper.find(".invalid-feedback");
+        expectTranslated(invalidFeedback,
+            "Please enter a valid email address",
+            "S'il vous plaît, mettez une adresse email valide",
+            "Por favor insira um endereço de e-mail válido", store)
+    });
+
+    it("does not validate email without @ symbol", async () => {
+        const store = createStore({}, {}, {}, true)
+        const wrapper = mount(ErrorReport, {
+            propsData: {
+                open: true
+            },
+            store
+        });
+
+        wrapper.find("#description").setValue("something");
+        wrapper.find("#reproduce").setValue("reproduce steps");
+        wrapper.find("#section").setValue("downloadResults");
+        wrapper.find("#email").setValue("testemail.com");
+
+        expect(wrapper.find(".btn-red").text()).toBe("Send");
+        expect(wrapper.find(".btn-red").attributes("disabled")).toBe("disabled")
+
+        const invalidFeedback = wrapper.find(".invalid-feedback");
+        expectTranslated(invalidFeedback,
+            "Please enter a valid email address",
+            "S'il vous plaît, mettez une adresse email valide",
+            "Por favor insira um endereço de e-mail válido", store)
+    });
+
+    it("disables button and display email validation message when invalid email address", async () => {
+        const store = createStore({}, {}, {}, true)
+        const wrapper = mount(ErrorReport, {
+            propsData: {
+                open: true
+            },
+            store
+        });
+
+        wrapper.find("#description").setValue("something");
+        wrapper.find("#reproduce").setValue("reproduce steps");
+        wrapper.find("#section").setValue("downloadResults");
+        wrapper.find("#email").setValue("test@email.c");
+
+        expect(wrapper.find("#email").attributes("pattern")).toBe(emailRegex.source)
+        expect(wrapper.find(".btn-red").text()).toBe("Send");
+        expect(wrapper.find(".btn-red").attributes("disabled")).toBe("disabled")
+
+        const invalidFeedback = wrapper.find(".invalid-feedback");
+        expectTranslated(invalidFeedback,
+            "Please enter a valid email address",
+            "S'il vous plaît, mettez une adresse email valide",
+            "Por favor insira um endereço de e-mail válido", store)
+    });
+
+
+    it("validates email and enables send button ", async () => {
+        const store = createStore({}, {}, {}, true)
+        const wrapper = mount(ErrorReport, {
+            propsData: {
+                open: true
+            },
+            store
+        });
+
+        wrapper.find("#description").setValue("something");
+        wrapper.find("#reproduce").setValue("reproduce steps");
+        wrapper.find("#section").setValue("downloadResults");
+        wrapper.find("#email").setValue("test@email.com");
+
+        expect(wrapper.find("#email").attributes("pattern")).toBe(emailRegex.source)
+        expect(wrapper.find(".btn-red").text()).toBe("Send");
+        expect(wrapper.find(".btn-red").attributes("disabled")).toBeUndefined()
+        expect(wrapper.find(".invalid-feedback").exists()).toBe(false)
     });
 
     it("invokes generateErrorReport action and sets showFeedback on send", async () => {